+++ jest.base.config.ts
@@ -1,6 +1,0 @@
-export const baseConfig = {
-    preset: "ts-jest",
-    transform: {
-      "^.+\\.ts$": "ts-jest"
-    }
-}