+++ packages/client/src/index.ts
@@ -1,4 +1,6 @@
 /* istanbul ignore file */
 
-export * from "./WalletAPIClient";
+export { default as HWTransport } from "@ledgerhq/hw-transport";
 export * from "@ledgerhq/wallet-api-core";
+export * from "./TransportWalletAPI";
+export * from "./WalletAPIClient";