+++ packages/server/src/internalHandlers/index.ts
@@ -2,8 +2,10 @@ import type { WalletHandlers } from "@ledgerhq/wallet-api-core";
 import type { TransformHandler } from "../types";
 import * as account from "./account";
 import * as currency from "./currency";
+import * as device from "./device";
 import * as message from "./message";
 import * as transaction from "./transaction";
+import * as wallet from "./wallet";
 
 type InternalHandlers = TransformHandler<WalletHandlers>;
 
@@ -14,8 +16,14 @@ export const internalHandlers: InternalHandlers = {
 
   "currency.list": currency.list,
 
+  "device.close": device.close,
+  "device.exchange": device.exchange,
+  "device.transport": device.transport,
+
   "message.sign": message.sign,
 
   "transaction.sign": transaction.sign,
   "transaction.signAndBroadcast": transaction.signAndBroadcast,
+
+  "wallet.capabilities": wallet.capabilities,
 };